+++ sdks/go/node/core/internal/fakes/caller.go
@@ -9,7 +9,7 @@ import (
 )
 
 type FakeCaller struct {
-	CallStub        func(context.Context, string, map[string]*model.Value, *model.CallSpec, string, *string, string)
+	CallStub        func(context.Context, string, map[string]*model.Value, *model.CallSpec, string, *string, string) (map[string]*model.Value, error)
 	callMutex       sync.RWMutex
 	callArgsForCall []struct {
 		arg1 context.Context
@@ -20,12 +20,21 @@ type FakeCaller struct {
 		arg6 *string
 		arg7 string
 	}
+	callReturns struct {
+		result1 map[string]*model.Value
+		result2 error
+	}
+	callReturnsOnCall map[int]struct {
+		result1 map[string]*model.Value
+		result2 error
+	}
 	invocations      map[string][][]interface{}
 	invocationsMutex sync.RWMutex
 }
 
-func (fake *FakeCaller) Call(arg1 context.Context, arg2 string, arg3 map[string]*model.Value, arg4 *model.CallSpec, arg5 string, arg6 *string, arg7 string) {
+func (fake *FakeCaller) Call(arg1 context.Context, arg2 string, arg3 map[string]*model.Value, arg4 *model.CallSpec, arg5 string, arg6 *string, arg7 string) (map[string]*model.Value, error) {
 	fake.callMutex.Lock()
+	ret, specificReturn := fake.callReturnsOnCall[len(fake.callArgsForCall)]
 	fake.callArgsForCall = append(fake.callArgsForCall, struct {
 		arg1 context.Context
 		arg2 string
@@ -38,8 +47,13 @@ func (fake *FakeCaller) Call(arg1 context.Context, arg2 string, arg3 map[string]
 	fake.recordInvocation("Call", []interface{}{arg1, arg2, arg3, arg4, arg5, arg6, arg7})
 	fake.callMutex.Unlock()
 	if fake.CallStub != nil {
-		fake.CallStub(arg1, arg2, arg3, arg4, arg5, arg6, arg7)
+		return fake.CallStub(arg1, arg2, arg3, arg4, arg5, arg6, arg7)
 	}
+	if specificReturn {
+		return ret.result1, ret.result2
+	}
+	fakeReturns := fake.callReturns
+	return fakeReturns.result1, fakeReturns.result2
 }
 
 func (fake *FakeCaller) CallCallCount() int {
@@ -48,7 +62,7 @@ func (fake *FakeCaller) CallCallCount() int {
 	return len(fake.callArgsForCall)
 }
 
-func (fake *FakeCaller) CallCalls(stub func(context.Context, string, map[string]*model.Value, *model.CallSpec, string, *string, string)) {
+func (fake *FakeCaller) CallCalls(stub func(context.Context, string, map[string]*model.Value, *model.CallSpec, string, *string, string) (map[string]*model.Value, error)) {
 	fake.callMutex.Lock()
 	defer fake.callMutex.Unlock()
 	fake.CallStub = stub
@@ -59,6 +73,32 @@ func (fake *FakeCaller) CallArgsForCall(i int) (context.Context, string, map[str
 	defer fake.callMutex.RUnlock()
 	argsForCall := fake.callArgsForCall[i]
 	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3, argsForCall.arg4, argsForCall.arg5, argsForCall.arg6, argsForCall.arg7
+}
+
+func (fake *FakeCaller) CallReturns(result1 map[string]*model.Value, result2 error) {
+	fake.callMutex.Lock()
+	defer fake.callMutex.Unlock()
+	fake.CallStub = nil
+	fake.callReturns = struct {
+		result1 map[string]*model.Value
+		result2 error
+	}{result1, result2}
+}
+
+func (fake *FakeCaller) CallReturnsOnCall(i int, result1 map[string]*model.Value, result2 error) {
+	fake.callMutex.Lock()
+	defer fake.callMutex.Unlock()
+	fake.CallStub = nil
+	if fake.callReturnsOnCall == nil {
+		fake.callReturnsOnCall = make(map[int]struct {
+			result1 map[string]*model.Value
+			result2 error
+		})
+	}
+	fake.callReturnsOnCall[i] = struct {
+		result1 map[string]*model.Value
+		result2 error
+	}{result1, result2}
 }
 
 func (fake *FakeCaller) Invocations() map[string][][]interface{} {