+++ sdks/go/node/core/internal/fakes/opCaller.go
@@ -9,7 +9,7 @@ import (
 )
 
 type FakeOpCaller struct {
-	CallStub        func(context.Context, *model.OpCall, map[string]*model.Value, *string, *model.OpCallSpec)
+	CallStub        func(context.Context, *model.OpCall, map[string]*model.Value, *string, *model.OpCallSpec) (map[string]*model.Value, error)
 	callMutex       sync.RWMutex
 	callArgsForCall []struct {
 		arg1 context.Context
@@ -18,12 +18,21 @@ type FakeOpCaller struct {
 		arg4 *string
 		arg5 *model.OpCallSpec
 	}
+	callReturns struct {
+		result1 map[string]*model.Value
+		result2 error
+	}
+	callReturnsOnCall map[int]struct {
+		result1 map[string]*model.Value
+		result2 error
+	}
 	invocations      map[string][][]interface{}
 	invocationsMutex sync.RWMutex
 }
 
-func (fake *FakeOpCaller) Call(arg1 context.Context, arg2 *model.OpCall, arg3 map[string]*model.Value, arg4 *string, arg5 *model.OpCallSpec) {
+func (fake *FakeOpCaller) Call(arg1 context.Context, arg2 *model.OpCall, arg3 map[string]*model.Value, arg4 *string, arg5 *model.OpCallSpec) (map[string]*model.Value, error) {
 	fake.callMutex.Lock()
+	ret, specificReturn := fake.callReturnsOnCall[len(fake.callArgsForCall)]
 	fake.callArgsForCall = append(fake.callArgsForCall, struct {
 		arg1 context.Context
 		arg2 *model.OpCall
@@ -34,8 +43,13 @@ func (fake *FakeOpCaller) Call(arg1 context.Context, arg2 *model.OpCall, arg3 ma
 	fake.recordInvocation("Call", []interface{}{arg1, arg2, arg3, arg4, arg5})
 	fake.callMutex.Unlock()
 	if fake.CallStub != nil {
-		fake.CallStub(arg1, arg2, arg3, arg4, arg5)
+		return fake.CallStub(arg1, arg2, arg3, arg4, arg5)
 	}
+	if specificReturn {
+		return ret.result1, ret.result2
+	}
+	fakeReturns := fake.callReturns
+	return fakeReturns.result1, fakeReturns.result2
 }
 
 func (fake *FakeOpCaller) CallCallCount() int {
@@ -44,7 +58,7 @@ func (fake *FakeOpCaller) CallCallCount() int {
 	return len(fake.callArgsForCall)
 }
 
-func (fake *FakeOpCaller) CallCalls(stub func(context.Context, *model.OpCall, map[string]*model.Value, *string, *model.OpCallSpec)) {
+func (fake *FakeOpCaller) CallCalls(stub func(context.Context, *model.OpCall, map[string]*model.Value, *string, *model.OpCallSpec) (map[string]*model.Value, error)) {
 	fake.callMutex.Lock()
 	defer fake.callMutex.Unlock()
 	fake.CallStub = stub
@@ -55,6 +69,32 @@ func (fake *FakeOpCaller) CallArgsForCall(i int) (context.Context, *model.OpCall
 	defer fake.callMutex.RUnlock()
 	argsForCall := fake.callArgsForCall[i]
 	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3, argsForCall.arg4, argsForCall.arg5
+}
+
+func (fake *FakeOpCaller) CallReturns(result1 map[string]*model.Value, result2 error) {
+	fake.callMutex.Lock()
+	defer fake.callMutex.Unlock()
+	fake.CallStub = nil
+	fake.callReturns = struct {
+		result1 map[string]*model.Value
+		result2 error
+	}{result1, result2}
+}
+
+func (fake *FakeOpCaller) CallReturnsOnCall(i int, result1 map[string]*model.Value, result2 error) {
+	fake.callMutex.Lock()
+	defer fake.callMutex.Unlock()
+	fake.CallStub = nil
+	if fake.callReturnsOnCall == nil {
+		fake.callReturnsOnCall = make(map[int]struct {
+			result1 map[string]*model.Value
+			result2 error
+		})
+	}
+	fake.callReturnsOnCall[i] = struct {
+		result1 map[string]*model.Value
+		result2 error
+	}{result1, result2}
 }
 
 func (fake *FakeOpCaller) Invocations() map[string][][]interface{} {